Kural uyumu: emoji ve ok karakterleri temizlendi; fiziksel baskı ön talebe çevrildi
CONTRIBUTING.md kendi kuralını söylüyordu (emoji ve ok hiçbir dosyada olmaz)
ama README, CHANGELOG, ERRATA ve bazı .md/.sql dosyaları bunları taşıyordu.

Fiziksel baskı: baskı henüz yapılmadı. SQL Server 2025 geliştirilmeye devam
eden bir sistem ve kitap her CU'da değişiyor; kart ön talep olarak yazıldı,
sayfa sayısı iddiası kaldırıldı.

+++ kod-ornekleri/bolum-22/06-embedding-refresh-strateji.sql
@@ -38,7 +38,7 @@ GO
 IF OBJECT_ID('ai.embedding_queue','U') IS NULL
 BEGIN
     CREATE TABLE ai.embedding_queue (
-        chunk_id BIGINT NOT NULL PRIMARY KEY REFERENCES ai.document_chunks(chunk_id),
+        chunk_id INT NOT NULL PRIMARY KEY REFERENCES ai.document_chunks(chunk_id),
         queued_at DATETIME2(0) NOT NULL DEFAULT SYSUTCDATETIME(),
         last_attempt DATETIME2(0) NULL,
         attempts INT NOT NULL DEFAULT 0,
